[yaml] Rename yaml to _yaml to prevent conflicts (#576)
[yaml] Rename yaml to _yaml to prevent conflicts

Change:

It was impossible to run ansible-navigator or ansible-lint or some
other things from within src/ansible_navigator due to the conflicting
yaml.py filename. This is just a simple rename since yaml is a
commonly used library.

Test Plan:

Ran ansible-navigator locally (with change from #575)


(cherry picked from commit 0918b3c142fb6dc5d1d1aa5f9704434397edf6f3)
