+++ src/ansible_navigator/actions/sample_form.py
@@ -6,7 +6,7 @@ from ..app_public import AppPublic
 from ..ui_framework import Interaction
 from ..ui_framework import dict_to_form
 from ..ui_framework import form_to_dict
-from ..yaml import yaml
+from .._yaml import yaml
 
 FORM = """
 form:
