+++ src/ansible_navigator/actions/write_file.py
@@ -10,7 +10,7 @@ from ..ui_framework import Interaction
 
 from ..utils import remove_dbl_un
 
-from ..yaml import human_dump
+from .._yaml import human_dump
 
 
 @actions.register
