+++ tests/unit/configuration_subsystem/conftest.py
@@ -18,8 +18,8 @@ from ansible_navigator.configuration_subsystem.navigator_configuration import Na
 from ansible_navigator.utils import ExitMessage
 from ansible_navigator.utils import LogMessage
 
-from ansible_navigator.yaml import yaml
-from ansible_navigator.yaml import Loader
+from ansible_navigator._yaml import yaml
+from ansible_navigator._yaml import Loader
 
 from ...defaults import FIXTURES_DIR
