+++ tests/unit/configuration_subsystem/test_fixture_sanity.py
@@ -4,8 +4,8 @@ entries for for all entries.
 import os
 
 from ansible_navigator.configuration_subsystem import NavigatorConfiguration
-from ansible_navigator.yaml import yaml
-from ansible_navigator.yaml import Loader
+from ansible_navigator._yaml import yaml
+from ansible_navigator._yaml import Loader
 
 from .data import ENVVAR_DATA
